+++ .gitignore
@@ -1,8 +1,9 @@
 __pycache__/
 env/
+venv/
 .idea/
 annotated_data/
-mibipy/__pycache__/
+src/__pycache__/
 dnn_vessel_heterogeneity/topic_generation/__pycache__/
 trained_models/
 runs/
