+++ src/data_analysis/umap_analyzer.py
@@ -13,12 +13,12 @@ import umap
 from sklearn.cluster import DBSCAN
 import seaborn as sns
 
-from mibipy.data_analysis.base_analyzer import BaseAnalyzer
-from mibipy.data_loading.mibi_data_feed import MIBIDataFeed
-from mibipy.data_loading.mibi_loader import MIBILoader
-from mibipy.data_loading.mibi_point_contours import MIBIPointContours
-from mibipy.data_preprocessing.markers_feature_gen import *
-from mibipy.plotting.visualizer import Visualizer
+from src.data_analysis.base_analyzer import BaseAnalyzer
+from src.data_loading.mibi_data_feed import MIBIDataFeed
+from src.data_loading.mibi_loader import MIBILoader
+from src.data_loading.mibi_point_contours import MIBIPointContours
+from src.data_preprocessing.markers_feature_gen import *
+from src.plotting.visualizer import Visualizer
 from config.config_settings import Config
 from src.data_analysis._cluster import k_means, dbscan, agglomerative, hdbscan_method
