+++ src/data_loading/mibi_loader.py
@@ -1,8 +1,8 @@
 from tqdm import tqdm
 
-from mibipy.data_loading.mibi_data_feed import MIBIDataFeed
-from mibipy.data_loading.mibi_reader import MIBIReader
-from mibipy.data_preprocessing.markers_feature_gen import *
+from src.data_loading.mibi_data_feed import MIBIDataFeed
+from src.data_loading.mibi_reader import MIBIReader
+from src.data_preprocessing.markers_feature_gen import *
 
 logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)
