+++ src/data_preprocessing/markers_feature_gen.py
@@ -11,7 +11,7 @@ import matplotlib.pyplot as plt
 from PIL import Image
 import pandas as pd
 
-from mibipy.utils.utils_functions import mkdir_p
+from src.utils.utils_functions import mkdir_p
 from config.config_settings import Config
 
 '''
