+++ src/mibi_pipeline.py
@@ -4,13 +4,13 @@ from multiprocessing import Pool
 
 from tqdm import tqdm
 
-from mibipy.data_analysis.base_analyzer import BaseAnalyzer
-from mibipy.data_analysis.vessel_asymmetry_analyzer import VesselAsymmetryAnalyzer
-from mibipy.data_loading.mibi_data_feed import MIBIDataFeed
-from mibipy.data_loading.mibi_loader import MIBILoader
-from mibipy.data_loading.mibi_point_contours import MIBIPointContours
-from mibipy.data_preprocessing.markers_feature_gen import *
-from mibipy.plotting.visualizer import Visualizer
+from src.data_analysis.base_analyzer import BaseAnalyzer
+from src.data_analysis.vessel_asymmetry_analyzer import VesselAsymmetryAnalyzer
+from src.data_loading.mibi_data_feed import MIBIDataFeed
+from src.data_loading.mibi_loader import MIBILoader
+from src.data_loading.mibi_point_contours import MIBIPointContours
+from src.data_preprocessing.markers_feature_gen import *
+from src.plotting.visualizer import Visualizer
 from config.config_settings import Config
 
 logging.basicConfig(format='%(levelname)s:%(message)s', level=logging.INFO)
